Support for updated query syntax.
This adds support for the new query syntax that is used on Citrination. This adds a PifSystemReturningQuery and DatasetReturningQuery object that share a common parent. This allows for a common interface to querying for PIFs or datasets by the content of either/both. The only difference between these two returning queries is the formatting that they allow that is specific to the type that each returns.

The set of features in this version matches exactly what is available in JCC 2.4.0, the current release version of that package.

All deployments of Citrination currently support the updated syntax. Servers are backwards compatible with v1.x and v2.x of PyCC, but v3.x adds significantly improved control on queries.