Added link to tutorials
One challenge for users who are trying to troubleshoot/or self-teach is that the tutorials do not render correctly when accessed from the docs/source/tutorials. Max G provided a link to an HTML version which displays correctly. You can't currently find this link from within the repo itself, which is why I am adding this.